requirements.txt
requests
neo4j
python-dotenv
networkx
pikepdf
progress
pdfminer.six # Debian requirement: sudo apt-get install build-essential python3-dev libffi-dev
